// crates/desktop/src/generated/dto/MarketMcpInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type MarketMcpInput = {
	id: string;
	label: string;
	default: string | null;
	placeholder: string | null;
	description: string | null;
	is_required: boolean;
	is_secret: boolean;
	format: string;
	choices: Array<string>;
};